requirements.txt
# AI CoinJoin Investigation System Requirements

# Core dependencies
requests>=2.28.0
numpy>=1.24.0
pandas>=2.0.0

# Configuration and data handling
PyYAML>=6.0
python-dateutil>=2.8.0

# Statistics and analysis
scipy>=1.10.0
statistics

# API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
aiohttp>=3.9.0

# Neo4j database
neo4j>=5.15.0

# Optional: Machine learning (for future enhancements)
# scikit-learn>=1.3.0
# xgboost>=1.7.0
# lightgbm>=4.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0

# Logging and monitoring
colorlog>=6.7.0

# Data validation
jsonschema>=4.17.0

# Performance optimization
psutil>=5.9.0

scikit-learn >= 1.3.0
xgboost >= 1.7.0